moved shared fixtures
